// frontend/src/app/shared/models/customer.ts
export class User {
    $key: string;
    userName: string;
    emailId: string;
    password: string;
    phoneNumber: string;
    createdOn: string;
    isAdmin: boolean;
  }
  
  export class CustomerDetail {
    $key: string;
    firstName: string;
    lastName: string;
    emailId: string;
    phoneNumber: string;
  }
  